# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: LabAutomation API
  version: 0.1.0
  description: >-
    Minimal начальный OpenAPI для системы автоматизированного оценивания.
servers:
  - url: https://api.example.com/v1
...
paths:
  /collections:
    get:
      summary: Список коллекций (лабораторные наборы)
      tags: [Collections]
      parameters:
        - in: query
          name: courseId
          schema:
            type: string
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Collections list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Collection'
                  total:
                    type: integer
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/ServerError' }
    post:
      summary: Создать коллекцию
      tags: [Collections]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collection'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/ServerError' }
  /collections/{collectionId}:
    get:
      summary: Получить коллекцию
      tags: [Collections]
      parameters:
        - name: collectionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Collection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collection'
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }
    patch:
      summary: Обновить коллекцию
      tags: [Collections]
      security:
        - bearerAuth: []
      parameters:
        - name: collectionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collection'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }
...
components:
  schemas:
...
    Error:
      type: object
      required: [error, code]
      properties:
        code:
          type: string
          description: Внутренний или стандартный код ошибки (ERR_*)
          example: ERR_NOT_FOUND
        error:
          type: string
          description: Короткое описание
          example: Resource not found
        details:
          type: object
          description: Дополнительные поля (контекст)
          additionalProperties: true
        traceId:
          type: string
          description: Корреляционный идентификатор трассировки
...
    Collection:
      type: object
      properties:
        id:
          type: string
        courseId:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        deadline:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [draft, published, archived]
        labsCount:
          type: integer
    CollectionCreate:
      type: object
      required: [courseId, name]
      properties:
        courseId:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        deadline:
          type: string
          format: date-time
          nullable: true
    CollectionUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        deadline:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [draft, published, archived]
...
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Требуется аутентификация
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServerError:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []